Linkedin job post finder/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
streamlit==1.29.0
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0 